refactor: SessionStatus에 수강 신청 가능 상태 홧린 로직 구현 및 테스트 수정
